Subnets
Consider the network id 192.168.4.0 or such relevant IP and create four subnets namely A, B, C, D. Assign the subnet mask. Write a Python \ C++ program to perform the following operations (use overloading if applicable). a) Ping the machine of same subnet. b) Ping the machine in subnet A from machine of subnet B. c) Analyze the output of the above sub assignments.